This version introduces a feature for alternative rendering based on filename extension requests. requests to /posts will use the templating engine requests to /posts.xml will display posts in xml form
Added public var "controller" to Application 
haxigniter.libraries.Request sets this variable

Added public var "format" to haxigniter.libraries.Application and haxigniter.libraries.Controller. it is a Hash<Dynamic> which is set in the app's controller's action method.
it's a list of formats that you want to allow requests to render as. The Key is the 'format', the Value is an iterable object.

haxigniter.libraries.Url.format is a new variable that returns the requested file extension, and defaults it to "html". lots of hooks default the 'format' to "html", which renders using the template engine normally.

added "renderFormat()" to ViewEngine to handle displaying of alternate formats.

added the first "Format Renderer" for "xml"
goal is to recursively iterate through each iterable and display meaningful xml data passed in from the controller.
currently tested to be working with the List<ResultSet> like:

public function index(){
	var posts = this.db.query('SELECT * FROM posts').results();
	this.view.assign('posts', posts);

	format.set('xml', posts);
	this.view.displayDefault();
}

public function show(){
	var post:ResultSet = this.db.queryRow('SELECT * FROM posts WHERE id=?',[id]);
	this.view.assign('post', post);
		
	var list:List<ResultSet> = new List<ResultSet>(); // cause xml renderer likes to start with an iterable
	list.add(post);
	format.set('xml', list);
	this.view.displayDefault();
}